edited types.ts
