+++ src/components/FavoriteButton/FavoriteButton.tsx
@@ -1,17 +1,15 @@
-import { StarIcon } from '../../assets/icons/StarIcon';
-import { StarIconActive } from '../../assets/icons/StarIconActive';
-import styles from './FavoriteButton.module.scss';
+import { StarIcon } from "../../assets/icons/StarIcon";
+import { StarIconActive } from "../../assets/icons/StarIconActive";
+import styles from "./FavoriteButton.module.scss";
+import { FavoriteButtonProps } from "../../types/types";
 
- type FavoriteButtonProps = {
-    isFavorite: boolean;
-    onClick: () => void;
-    className?: string; 
-  }
-
-export const FavoriteButton = ({ isFavorite, onClick }: FavoriteButtonProps) => {
-  return  (
+export const FavoriteButton = ({
+  isFavorite,
+  onClick,
+}: FavoriteButtonProps) => {
+  return (
     <button className={styles.starIcon} onClick={onClick}>
-       { isFavorite  ? <StarIconActive /> : <StarIcon /> }
+      {isFavorite ? <StarIconActive /> : <StarIcon />}
     </button>
-  )
+  );
 };
